Experiments into Big Endian and unaligned access support.
This is based on https://github.com/azlicic/htslib too, but tidied up
with use of dedicated memory accessing functions.

TODO: The cram code is still broken. It passes self tests, but by
virtue of not being compatible. (It has an implicit assumption that
the auxiliary string is the same in memory as on disk, which is not
true for HTSlib.)