Ditched the mFILE.c in favour of hfile mmap interface.
We also use the preload: scheme on non-local files (eg urls) so that
we can operate in the same manner whether the file was mmapped or read
into memory.

Also fixed a bug in hfile_mem_get_buffer which returned the wrong
length.

Finally (not directly connected with this), the Makefile gained a
"make relink" target which is useful for linking against new
dependencies or third party libs without a full clean and rebuild.