Correction to mmap hfile code.
The code only does mmap on files >= 1024 bytes long.  I'm not sure if
this is ideal, but it does mean we cannot just assume we'll get the
buffer back out of it.  Instead we now cope with preload:file combined
with mmap mode to ensure one way or another the full contents are
known after opening.