Fix crash when Windows external plug-ins register handlers
Windows dlls resolve symbols at link time (unlike UNIX shared
objects which resolve at run time).  This means Windows plug-in
dlls are linked against and use the HTSlib shared library even if
the executable that loads them statically links libhts.a.  This
caused plug-ins loaded by statically linked executables to crash
when they called hfile_add_scheme_handler as the schemes khash in
the dll had not been set up correctly (and even if it had been,
the executable would still not be aware of the scheme handler).

Try to fix by spotting when schemes is NULL, and if so try to
find (via dlopen, dlsym) the executable's copy of
hfile_add_scheme_handler and call that instead.  If that doesn't
work, print a warning and carry on, which is better than outright
crashing.