Further speed up of VCF parser (formats).
The if elseif checks are now a switch statement and juggled in order a
bit.  Also the fmt[j].max_m type code is now f->max_m with f
incremented along with j.

The impact on gcc builds is minor (maybe 1%), but for clang this was
8-9% speed improvement on a 1000 genome multi-sample VCF.  Example
timings for clang:

Previous commit
      23333.77 msec task-clock                #    1.000 CPUs utilized
   83667512727      cycles                    #    3.586 GHz
  199145555089      instructions              #    2.38  insn per cycle
   43099743981      branches                  # 1847.097 M/sec
     665687093      branch-misses             #    1.54% of all branches

This commit
   75967289857      cycles                    #    3.585 GHz
  195076309580      instructions              #    2.57  insn per cycle
   43265084488      branches                  # 2041.736 M/sec
     640008186      branch-misses             #    1.48% of all branches