A couple more fixes to the VCF-as-text optimisation.
However this now causes a test failure.  Specifically

    ./htsfile -c test/noroundtrip.vcf

produces FORMAT of "GT:GT 0/1" which is identical to the input data,
but if we force into BCF and back out again we get "GT:GT 2,4:." and
this is what the test is expecting to happen.

I'm not sure this is something we can fix.  Bad input is just
regurgiated back, but see 8f782d12 for the source of this test.
Is the test useful to keep?