Use HI (hit index) aux tag when merging split reads.
Multiple supplementaries or multiple secondaries means we cannot rely
purely on flag+name as a disambiguation.  Once split, we must know
which splits belong together, so we use the hit-index for that.

This needs defining in the spec: HI is mandatory for non-primary
split-reads.

TODO:  the splitting side needs to automatically add this.