Fixed merging via adding new HS and HE tags.
We couldn't rely on cigar and H tag as this is misleading.  It shows
the number of seq bases, not ref bases.  For that we need explicit
tags.  We have hit-start and hit-end tags so each sub-read explicitly
lists the full reference dimensions of the full read.  We then use
that in the merge instead of trying to be clever.