Next update for CRAM 4.0; now matching io_lib.
This is the htslib equivalent to io_lib's 090dfd3f and d86b008.
It is now in adherance to the draft CRAM 4 spec.

It removes the E_SINT and E_SLONG types and replaces E_EXTERNAL with
E_BYTE, E_VARINT_SIGNED and E_VARINT_UNSIGNED.  Integer data types
aren't explicitly signed or unsigned but simply use the codec that
works for them.  So if all ref-ids are >= 0 it'll be VARINT_UNSIGNED,
otherwise VARINT_SIGNED.  This future proofs things by adding
introspection to the format.

Also removed HUFFMAN from CRAM 4.  It was only being used for storing
constant values, so we now have CONST_BYTE and CONST_INT instead
(which always uses a signed encoding).

This should be the final removal of ever encoding negatives as if they
were positives (eg -1 as 0xFFFFFFFF) and so in turn removes all size
limitations on values.