Fix "make check" on MinGW.
If we haven't previously done a "make" or "make all", the .dll file
won't have been created.  The test/with-shlib.sh script will then fail
as the "cp -p ../../hts-*.dll ." command fails.  We no longer even
attempt to execute this test when plugins don't exist (which they
won't on Windows unless we've installed the dlfcn package).

Note the analogous non-mingw commands are "ln -s ../../libhts.so.* ."
and similar, which don't fail in the $? sense, but do fail with
regards to produced a bogus target containing a wild-card pattern.

This implies the error checking in this script is broken for all
platforms bar MinGW, but given the very limited use for this one test
case it's an irrelevance.

The change also adds lib-shared to the plugins dependencies.  This
typically doesn't matter, but on Windows with different import and
export library filenames this is needed to build all the dependencies
required in order to get the tests working.

Finally, added mingw-w64-x86_64-dlfcn to recommended list of pacman
dependencies so we can use "./configure --enable-plugins" on MinGW if
we wish.