First draft of bgzip2 using zstd.
This has a larger default block size, which can also be adjusted.
It's a self-indexing file in the seekable-zstd format, meaning we can
do random access on the data using original file offsets.

This index is secondary to the genomic index, which is still required
for chromosome:start-end based queries.

We also have pzstd skippable frames preceding each data frame to hold
the compressed size of the next frame.  This permits easy
multi-threading capabilities including decoding.