Increase the input block size for bgzip.
Commit e495718 changed bgzip from unix raw POSIX read() calls to
hread().  Unfortunately hread gets its buffer size from stat of the
input file descriptor, which can be 4kb for a pipe.  We're reading
0xff00 bytes, so this ends up being split over two reads mostly, with
one or both involving additional memcpys.  This makes the buffered I/O
worse performing than non-buffered.  In the most extreme cases (cat
data | bgzip -l0 > /dev/null) this is a two fold slow down.

The easy solution is just to increase the buffer size to something
sensible.  It's a little messy as we have to use hfile_internal.h to
get hfile_set_blksize, but it works.  I'm not sure why we didn't elect
to make that API more public.  Probably simply out of caution.

Fixes #1767