Make test_view handle region queries on VCF.gz
Why is this so rediculously complicated?  I think it says something
about our myriad of indexing formats if a csi produced for VCF and a
csi produced for BCF requires a totally different API for querying.

It appears the only difference between the two is the anonymous (and
undocumented) meta-data block of CSI, which for BCF looks to hold data
for all chromosomes in the header while for VCF holds the data only
for chromosomes present.  Hence foo.chr20.vcf.gz with one chromosome
but a header full of them needs tid=0 for vcf and tid=19 for bcf.
Beyond that I think it ought to largely be the same interface.
Perhaps I'm missing something though?

Note this is not even *tabix* tbi either, but CSI in both cases.
I'm confused how we made it so hard.