Fix memory leak when writing threaded VCF.gz.
The call to hts_set_opt(out, HTS_OPT_THREAD_POOL, &p) in test_view
calls sam_state_create because that code does this for both SAM and
text_format, so the multi-threaded SAM can work irrespective of
whether we've written the header yet.

Unfortunately the same logic means a newly opened VCF file without
header yet written is also thinking it's SAM.gz.  This shows as a
memory leak, and was present in develop too.