Fix a bug with too small a block size breaking the index.
This shouldn't happen anyway as ideally we don't split a record
between blocks (this will be a separate fix), but *if* we do we should
be capable of correctly indexing the starting block still.

Also fix the buffer size as we grow it (eg for header) but we need to
shrink it back again so one large block doesn't erroneously mean many
large blocks from then on.