Stop large records from being split over multiple bgzf2 blocks.
We already use bgzf_flush_try to cause a flush if the record doesn't
fit in the remainder of the block, but we don't prevent the record
from being written out in multiple blocks if it's bigger than the
block size.

While bgzf2_write has a "can_split" parameter which informs us whether
we are permitted to split this write over multiple blocks or not, we
often write data in multiple bgzf2_write calls.  We have the dilemma
of not wanting a never ending series of can_split=0 write calls from
never flushing, vs not wanting to split multi-write records up.

Fortunately we already know how big the record will be as we called
bgzf_flush_try before to check, so we also use this call as a hint to
grow the next buffer size and ensure it never overflows the buffer.